Improve explaination for prefixes and find command in User Guide
There is an overlapping usage for the term "parameter" which currently refers to both the prefixes "e/, n/, etc." which can cause confusion for users. The instructions for the find command are particularly confusing as it requires dynamic usage of prefixes.

Let's update the user guide to be clearer in explaining the usage of prefixes and parameters.