Update User Stories
The user stories does not align with our overall
objective of creating a mini-CRM system.

This may make the overall project very confusing
if other developers were to see the developer
guide and not see the updated user stories.

Update the user stories to match what my group
is trying to attain.

This will be more accurate to what we would like
to achieve the next few weeks, and also to meet
the requirements of this module.

No other reference.