Update the developer guide
Update the DeveloperGuide.md file to update user stories, use cases and glossary.

The user stories are now more organized and more cohesive with the phrasing. Additional user stories were also added or removed.

The use cases are now more general to account for the similar uses for adding/removing/modifying of different types of information.

The glossary had some unnecessary terms and was missing the term 'tooltip'.