Change command format to "/name" style
For better clarity of prefixes and more natural UX, the fields are
mentioned in full such as "/name" instead of "/n" and the slash is
at the front of a field instead of being at the end.

No change to existing fields used.

Error message for duplicate prefixes changed to follow natural ordering
such that the prefixes displayed is consistent.