fix: change unique parameter of persons in the addressbook to email
initially the unique parameter of persons in the addressbook was the
name of the person. however, it does not make sense for the name of
the person to be unique as it is not uncommon for 2 persons in the
cohort to have the same name. therefore, we changed the unique parameter
to the email address of persons as 2 students in a cohort should not
have the same email address.

change,
* method in Person class that compares two persons if they are equal
to compare the email address of the persons instead of the name
* error message of the addressbook to warn users of a duplicate email
address in a contact they attempt to add