Add Copy Command
The only way to obtain the emails of people in the list is to manually
copy them one by one.

This is not ideal as a large list would take a proportionately long
time to obtain all the emails.

As such, lets implement the copy command to follow the behavior
suggested in the developer guide, copying all emails currently listed
in the last shown list. The command would take into consideration if the
latest list is empty, and notifies the user if there are no emails to
be copied. If the list is empty, the emails of all persons displayed
would be copied into the user's clipboard. The user can then paste this
anywhere using `ctrl+v`.

This command has been implemented to meet the identified needs of our
target users to aid in quickening communication between the user and
target groups.

This commit will also include all relevant tests to ensure that
the methods involved have been thoroughly tested.