Bug fix: Run in headless mode
The key reason why the test case fails the GitHub checks is because
the copy command relies on copying information to a keyboard. Since the
checks are done in headless mode, java.awt.HeadlessException will be
thrown. Hence the copy command will only copy to the clipboard if
it is in headed mode to avoid this issue.