Refactor addressbook into TaskMasterPro
There are some parts which I did not refactor. For example,
acknowledgements to the original addressbook project and
the content in the logs.

The other parts are mostly refactored but since there are
quite a number of files involved, do look out for potential
errors or portions which should not have been refactored.