Updated User Guide to reflect our planned v1.2 features.
Did not include the assign priority feature and add company tag feature.
I was not sure about how to go about it since both features are related
to the add command. To be further discussed and planned.

User Guide to be edited in the future when implementation of v1.2
features is completed to ensure an accurate User Guide.