Add command history feature
When entering a command, the up and down arrow keys can be used to view
previously entered commands. Additionally, any change to a command is
saved when navigating to another command, so that the user is able to
navigate to a previous command to refer to it, and return back to their
new command to finish it.