fix: update user guide to include more specific information on using the export command
this update ensure the user is fully aware of how the export command
works with the addressbook.

change the command word for exporting data to csv to `export`.
add information about the exported csv file and how it overwrites the
current file everytime the `export` command is ran if a csv file exists.