Add Studio parameter
Currently, the application is only able keep track of student classes using the "tag" functionality.

To allow for more restrictions to be placed on the structure of studio parameters and to pave the way for the implementation of future features, lets add an additional optional parameter of "Studio".

Code coverage was also added for the studio parameter, along with the matric parameter as well.