Create framework for Add Appointment
Create a number of changes to support adding
of appointments to MediCLI

 A typical patient management system would want
to track the appointments that the patients have.

The appointment class establishes an appointment
in a very basic sense which is between a doctor and
a patient at a particular date.

This allows for the very basic tracking of appointments.