feat: tighten constraints for Matric class.
Now, only matriculation numbers that start with 'A' are accepted.
This reflects in the lore of the National University of
Singapore when it was created on 8 August 1980, where it was
determined that 'A' would stand for 'Academic Year'. Hence there
is no reason to accept other characters for the first letter.

Lets take a nap