Update developer guide
Use cases are enumerated for easier inclusions in other use cases.

Inclusions are utilised in the use cases for pairing and tagging
contacts to minimise low-level steps. The use case for listing all
contacts is also used in many other use cases, includign pairing and
tagging.

Specific details are added to error messages in the use cases for pairing
and tagging to better understand what the error is.