Add untag command
As a counterpart to the tagging command, this command untags a user from
some tags.

This implementation can do multiple tags at the same time to amke
removing tags faster. It throws an error on non matching tags, but that
can be changed if that is not desirable e.g. if batch operations are
introduced.