Add appointment ui
Add UI to manage appointments

MediCLI needs to be able to display appointments
as they are added. Presently there is no way to do so.

This commit adds a UI that displays appointments  in a
list beside the person cards. This allows the user to see
all appointments and persons in one go, making it
convenient.