Person: Add reflection parameter
Currently, we have most of the student information such as name, email,
number and matriculation information, which is sufficient for
broadcasting messages to a large number if people.

However, if we want only want to contact specific reflection classes for
reasons such as class rescheduling, or other targeted announcements or
communication, it would be time consuming to contact each individual
student.

Let's create a reflection class parameter to associate each student with
a recitation class.

In doing so, we can use the find and copy functions to quickly contact
the students in specific reflection classes.

Unit testing has also been completed for new and edited code.