Add ToggleDisplayCommand functionality
Changed the ToggleCommand to ToggleDisplayCommand, to reflect the actual action performed more accurately. Edited the AddressBookParser to account for the new command.
Changed the CommandResult constructor, and added a new boolean parameter toggleDisplay to the constructor
Changed the constructors for the HelpCommand and the ExitCommand, which used the old constructor for CommandResult.
Added the toggleDisplay method into the MainWindow class.
Added the actual toggle display functionality and a method to the PersonListPanel class to actually toggle the display.
Created a new test class for the new command.
Edited all existing commands utilising the old constructor, and updated them to use the new constructor.