Remove Command: add enhanced version of Delete Command
The current Delete Command is inconvenient as it requires users to find
the index of the contact to delete, which is unscalable when the
address book expands. It is also unsafe, as there are no intermediate
steps to confirm the deletion.

The logic of (1) `remove <NAME>`, then listing an intermediate filtered
list of contacts with matching names, before (2) `remove <INDEX>` from
the much shorter list of matching contacts, is used.

Entering the name of the contact to remove is preferable over directly
using the index in large address books, where finding the index might
be inconvenient.

Presenting a shorter list of contacts with names matching the predicate
given in (1) then allows the index in (2) to be found more conveniently,
and serves as an intermediate step to help users confirm their deletion.