Fix .contains() method bug
The contains method for UniqueInternshipList is used to check
for unique internships. Internship has some identity fields,
namely companyName, location, role and description.

Currently, the .contains() method actually uses the stronger
definition of equality defined in Internship, which checks for
the equality of all data fields.

Let's change .contains() to use .isSameInternship() in Internship
that only checks for equality for the identity fields.