Set result display box to wrap result text
Currently the text overflows and expands the text area, so
the user is unable to see the full error/result message
unless he scrolls to the right with a mouse.

In light of efficiency purposes, it is better to have the text
wrap around the text area so it will not extend horizontally
past the border.

Let's set the wrap text property to true for better user QOL.