feat: update UML sequence diagram in developer guide to show sequence for export command
the sequence diagram is useful in showing how the components of the
address book interact with each other to respond to a user command. in
this case, the user command we are showing is `export`.

changes made include,
* sequence diagram for `delete` command has been updated to a sequence
diagram for `export` command, which is a new feature we included in
out v1.2
* developer guide has been updated to change any information related to
previous sequence diagram to be relevant to the current updated
sequence diagram

other note,
* imported packages and their classes which are used in ExportCommand
were not included in the sequence diagram. Some input on whether it is
relevant and should be added into the diagram would be appreciated :)