Update Model Diagram
The current diagram edited displays the various parameters associated
with the person class.

However, this diagram has become updated with the inclusion of new and
important parameters that should be displayed in the diagram.

Let's add the Matric, Studio and Reflection parameters into the diagram.

In doing so, we can make the developer guide more clear for future
developers who wish to work on this software.