Modify deleteByName method to handle multiple employees with same name
The deleteByName method in DeleteCommand.java previously deleted an
employee based on the name without checking if there were multiple
employees with the same name. This could lead to unintended deletions
if multiple employees shared a name.

This commit modifies the deleteByName method to first check if there
are multiple employees with the same name. If there are, it throws a
CommandException with a message asking the user to delete by uid. If
there is only one employee with the name, it deletes the employee.

This change improves the robustness of the delete command and prevents
unintended deletions.