feat: Introduce add_module command
This `add_module` command introduces new prefixes, new classes (i.e `AddStudentModuleCommand/Parser`, etc.).

In the parser layer, I've added a `AddStudentModuleCommandParser` that craps out a `AddStudentModuleCommand`. It then executes and deals with the `Student` model.

model: `Student` now supports a list of `modules` which are `ModuleCode`.

For now, we are storing a list of `ModuleCode` instead of `Module` in each Student model. This is because the ability to synchronize the state of Modules in `ModuleMap` and `Student` will be tough to keep track off, thus I've resorted to using `ModuleCode` instead.

For the storage layer, I've also added a `JsonAdaptedModuleCode`. I had to add this because there was some finicky issue with saving a List<String> within a `.json` file so I created the adapter and it fixed the issue. If anyone knows how to do it without creating that adapter feel free to correct this!