Implement deleteByUid feature in DeleteCommand and DeleteCommandParser
The DeleteCommand previously only supported deletion of employees by
index or name. However, this could lead to unintended deletions if
multiple employees shared a name.

This commit adds a new feature to the DeleteCommand that allows
users to delete employees by their unique id (uid). This is done
by adding a new constructor to the DeleteCommand that takes a uid,
and modifying the execute method to handle the case where a uid
is provided.

The DeleteCommandParser is also updated to parse the uid input
from the user. This involves checking if the input contains
"uid", and if so, creating a new DeleteCommand with the uid.

This change provides users with a more precise way to
delete employees, especially in cases where multiple
employees share a name.