feat: update export feature to export the listed or filtered person.
for ease of use of the application, users should be able to directly
export the filtered persons from the application instead of having to
additionally filter the contacts in the csv file exported when all
persons are exported.

add feature that only exports persons in the filteredPersonList of the
addressbook.