docs: update user guide to match new behaviour of the export command
previously the export command exports all data in the addressbook.
however, it is now changed to export only the filtered data. hence, the
instructions in the user guide under the export command should be
changed to match this new feature implemented.

include steps to guide users to export filtered data.