Improve UI
Currently, the UI feels messy and cluttered. the Studio and Reflection classes are stacked on top of all other information and the information is all stacked on top of one another.
Lets seperate out the classes (studio and reflection) and create a new UI space for them.
Lets also increase the space between information so it feels less cluttered