Hide empty Matric Value
Currently, when the matriculation value is empty, the text
"No matriculation number" is displayed in the person card in the UI.

However, this may cause the user to perceive this as an issue rather
than a feature since this list may contain staff who do not have
matriculation numbers.

Let's hide the matriculation number if it is empty.

As a result, each entry will only show the information available, which
makes reading data from the app more concise.