Bug fix: load issue when optional params missing
Currently, when either Matric, Studio or Reflection fields were empty
i.e. "", the code will throw an `IllegalValueException`.

This behavior is incorrect because these fields are optional and should
not need to be filled.

Let's ensure that the saved data is loaded correctly when the optional
parameters are empty whilst ensuring that the code ensures only valid
values are allowed.

Doing so will ensure that the code will achieve the correct behavior,
only throwing an `IllegalValueException` when it is appropriate.