Hide empty Matric Value
Currently, when the matriculation value is empty, the no text is
displayed in the person card in the UI, but the non-visible text
still takes space in the person card.

However, this may cause the user to perceive this as an issue rather
than a feature since there is a vacant space.

Let's remove the matriculation number component if it is empty.

As a result, each entry will be correctly sized whilst showing the
information available, which makes the experience of reading data from
the app more polished.