Update AssignedTasks
The storage now uses the hashtable on top of the existing String

The String is still necessary for transition between loading and storing
into the JSON file. Problem being that we cannot initiate the hashtable
until all Employees and Tasks has been created and added to
TaskMasterPro, so we need the String as the placeholder.

All functions aside from loading and saving should make use of this
hashtable for ease.

UI has been updated for the tasks assigned to not just show the String
tasks, but to iterate through the hashtable and display all tasks with
their ids and names.

The storage files were also updated to facilitate the initialisation of
the hashtable when loading files from the JSON file.