find: include new prefixes
Currently, contacts can now have newly added parameters with respective
prefixes for matric number, reflection and studio. However, the
predicates that are used by the find command do not support the new
optional parameters yet.

This needs to change because the basis of the find command is to make
communication between the head TA and the different groups of students
taking the course more streamlined, which means it would be more
meaningful to find contacts by the new parameters added.

Let's change the find command and the related classes to support
finding with the new optional parameters, implementing this in the same
way that the supported parameters are currently implemented.

This will allow for the functionality to better suit the intended
user, but will also maintain the current structure of the code,
which is beneficial for future developers.