Update UserGuide.md
Currently, the user guide has the generic introduction for AB3, and has
a section irrelevant to the current direction of the Avengers Assemble
application.

However, we want the user guide to be slightly more specific to better
suit the app, and remove unnecessary parts for the guide to be
completely relevant.

Let's touch up the introduction and remove the section on "archiving"
data files.

By improving the user guide, we can make the user experience more
straightforward.

Additional info:
The lines that say "You can refer to the [command format summary]
(#command-format-summary) for more guidelines for each parameter." have
been replaced with "For more details on each parameter,
[click here](#command-format-summary)." to make the user experience
more straightforward.