Update the DeveloperGuide for Logic
The DeveloperGuide is not up to date with our
implementations.

This may be confusing to other developers or
readers that are viewing it.

Update the DeveloperGuide to our implementation
logic and methods.

This will allow our DeveloperGuide to be up to
date with our changes and easier to understand
overall, allowing for easier further
implementations.