Find Command: adapt to work with scores
Currently, the find command can be used to group students by various
parameters such as name, email, phone etc.

However, we want to improve the functionality of the find command to
display students based on how they perform in assessments such as
exams.

Let's modify the find command to work with the Exam and Score classes
to improve the utility of the application for head TAs.

This change is crucial because it allows the head TA to better interact
with students based on how they do for exams. I.e. head TAs will
be able to contact students who underperform for remedials etc. or
reach out to top performers to gain insights on how they reach
a high proficiency in the tested topics.

Unit Testing and reformatting according to checkstyle will be done in
the next commit following this.