Implement `alist` that displays archived contacts
This was done by making sure the filteredPersons variable in ModelManager now correctly filters the json file to only contain either unarchived persons and archived persons. In theory, this should allow all commands to work on the archived list too.

Note that we have changed the ModelManager to show only the unarchived persons right from the start. If we showed all persons, it would also show the archived persons.

This currently works for find, filter and delete. Add and edit do not work yet.