Remove extra Constructor in Person Class
There is an extra constructor in the person class.
This is confusing to developers on which to use.

This could cause null pointer errors when the
wrong constructor is used, as the order field is
not initialized.

Remove the extra constructor and fixed codes that
call on it.

This adjustment will enable us to adhere to a
single constructor, facilitating a more
streamlined implementation.