Update DG to include copy implementation
Currently, the user guide has the implementations of planned
features.

However, it may not be sufficient for incoming developers who wish to
understand the implementation of the copy command.

Let's add in the implementation details for the copy command, with a
detailed explanation of the command, its considerations,
as well as an activity diagram showcasing the flow of the command.

By explaining the implementation of the copy in great detail, software
developers who wish to add onto the copy command will have a decent
understanding of the command, and it may help in the future development
of the application.